Architecture Proposal
I have created a series of directories for scripts, each folder has a blank.zs file in it to force github to recognize the folder. This also has the benefit of being a blank file ready to write script into (with a rename) for each category. We can delete directories we do not use when we have completed the pack. Not all possible methods may be yet included, more may come. Vanilla methods used will be included in the root folder of the given discipline rather than their own "Vanilla" folder at any level.